+++ crates/shared/primitives/src/test_utils/mod.rs
@@ -11,3 +11,6 @@ pub use contracts::{
     AccessListContract, ContractFactory, DoubleCounter, Logic, Logic2, Minimal7702Account,
     MockERC20, Proxy, SimpleStorage, TransparentUpgradeableProxy,
 };
+
+mod transactions;
+pub use transactions::build_eip1559_tx;
